Rename mapper methods and class name follow by PR advice
