+++ src/main/java/pl/ateam/disasteralerts/disasteralert/DisasterService.java
@@ -1,26 +1,13 @@
 package pl.ateam.disasteralerts.disasteralert;
 
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
+import jakarta.validation.Valid;
+import jakarta.validation.constraints.NotNull;
+import org.springframework.validation.annotation.Validated;
 import pl.ateam.disasteralerts.disasteralert.dto.DisasterAddDTO;
 import pl.ateam.disasteralerts.disasteralert.dto.DisasterDTO;
 
-@Service
-@RequiredArgsConstructor
-class DisasterService implements DisasterServiceAPI{
+@Validated
+public interface DisasterService {
 
-    private final DisasterRepository repository;
-    private final DisasterMapper mapper;
-
-    @Transactional
-    public DisasterDTO addDisaster(DisasterAddDTO disasterAddDTO){
-
-        Disaster disaster = mapper.disasterAddDtoToDisaster(disasterAddDTO);
-
-        DisasterDTO disasterDTO = mapper.disasterToDto(
-                repository.save(disaster));
-
-        return disasterDTO;
-    }
+    DisasterDTO addDisaster(@NotNull @Valid DisasterAddDTO disasterAddDTO);
 }
