--- src/main/java/pl/ateam/disasteralerts/disasteralert/DisasterServiceAPI.java
@@ -1,13 +1,0 @@
-package pl.ateam.disasteralerts.disasteralert;
-
-import jakarta.validation.Valid;
-import jakarta.validation.constraints.NotNull;
-import org.springframework.validation.annotation.Validated;
-import pl.ateam.disasteralerts.disasteralert.dto.DisasterAddDTO;
-import pl.ateam.disasteralerts.disasteralert.dto.DisasterDTO;
-
-@Validated
-public interface DisasterServiceAPI {
-
-    DisasterDTO addDisaster(@NotNull @Valid DisasterAddDTO disasterAddDTO);
-}
