+++ src/test/java/pl/ateam/disasteralerts/DisasterAlertsApplicationTests.java
@@ -8,8 +8,8 @@ import org.springframework.context.annotation.Import;
 @SpringBootTest
 class DisasterAlertsApplicationTests {
 
-    @Test
-    void contextLoads() {
-    }
+//    @Test
+//    void contextLoads() {
+//    }
 
 }
